# Conflicts:
#	.idea/modules.xml
#	app/src/main/java/com/example/parkly/Fragment/HomeFragment.java

+++ app/src/main/java/com/example/parkly/Fragment/HomeFragment.java
@@ -3,11 +3,13 @@ package com.example.parkly.Fragment;
 import android.content.Context;
 import android.content.DialogInterface;
 import android.content.Intent;
+import android.os.AsyncTask;
 import android.os.Bundle;
 import android.os.CountDownTimer;
 import android.support.annotation.Nullable;
 import android.support.v4.app.Fragment;
 import android.support.v7.app.AlertDialog;
+import android.support.v7.widget.AppCompatSpinner;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
@@ -15,16 +17,21 @@ import android.widget.AdapterView;
 import android.widget.ArrayAdapter;
 import android.widget.ListView;
 import android.widget.Spinner;
+import android.widget.SpinnerAdapter;
 import android.widget.TextView;
 import android.widget.Toast;
 
 import com.example.parkly.Activity.MainActivity;
 import com.example.parkly.Activity.addCarActivity;
+import com.example.parkly.DataBase.LicensePlateAdapter;
 import com.example.parkly.DataBase.Tables.LicensePlate;
 import com.example.parkly.DataBase.LicensePlateDatabase;
 import com.example.parkly.DataBase.LicensePlateRepository;
 import com.example.parkly.DataBase.LocalUserDataSource;
+import com.example.parkly.DataBase.Tables.LicensePlate;
 import com.example.parkly.R;
+
+import org.w3c.dom.Text;
 
 import java.io.BufferedReader;
 import java.io.File;
@@ -33,11 +40,15 @@ import java.io.FileNotFoundException;
 import java.io.FileOutputStream;
 import java.io.IOException;
 import java.io.InputStreamReader;
+import java.nio.Buffer;
+import java.nio.file.Files;
+import java.text.SimpleDateFormat;
 import java.util.ArrayList;
 import java.util.Calendar;
 import java.util.Date;
 import java.util.GregorianCalendar;
 import java.util.List;
+import java.util.Observable;
 import java.util.Scanner;
 
 import io.reactivex.ObservableEmitter;
@@ -45,6 +56,7 @@ import io.reactivex.ObservableOnSubscribe;
 import io.reactivex.android.schedulers.AndroidSchedulers;
 import io.reactivex.disposables.CompositeDisposable;
 import io.reactivex.disposables.Disposable;
+import io.reactivex.functions.Action;
 import io.reactivex.functions.Consumer;
 import io.reactivex.schedulers.Schedulers;
 
@@ -72,6 +84,7 @@ public class HomeFragment extends Fragment {
     private TextView ends;
     private TextView timeEnds;
     private TextView confirm;
+    private File file;
 
     //Adapter
     private Spinner spin_DefaultCar;
@@ -147,17 +160,17 @@ public class HomeFragment extends Fragment {
         final ArrayList<String> time = new ArrayList<String>();
 
 
-        listZones = (ListView)view.findViewById(R.id.list_zones);
-        listTime = (ListView)view.findViewById(R.id.list_time);
+        listZones = view.findViewById(R.id.list_zones);
+        listTime = view.findViewById(R.id.list_time);
         final ArrayAdapter<String> zonesAdapter = new ArrayAdapter<String>(getActivity(), android.R.layout.simple_list_item_single_choice, zones);
         final ArrayAdapter<String> timeAdapter = new ArrayAdapter<String>(getActivity(), android.R.layout.simple_list_item_single_choice, time);
         listZones.setAdapter(zonesAdapter);
         listTime.setAdapter(timeAdapter);
 
 
-        TextView outputPrice = (TextView)view.findViewById(R.id.txt_outputPrice);
+        TextView outputPrice = view.findViewById(R.id.txt_outputPrice);
         tempPrice = outputPrice;
-        TextView outputTime = (TextView)view.findViewById(R.id.txt_outputTime);
+        TextView outputTime = view.findViewById(R.id.txt_outputTime);
         tempTime = outputTime;
 
         disableEnableConfirm();
@@ -388,7 +401,7 @@ public class HomeFragment extends Fragment {
 
         //init View
         adapter = new ArrayAdapter<String>(getActivity(), android.R.layout.simple_spinner_dropdown_item, licensePlateList);
-        spin_DefaultCar = (Spinner) view.findViewById(R.id.spin_DefaultCar);
+        spin_DefaultCar = view.findViewById(R.id.spin_DefaultCar);
         registerForContextMenu(spin_DefaultCar);
         spin_DefaultCar.setAdapter(adapter);
     }
@@ -533,7 +546,7 @@ public class HomeFragment extends Fragment {
         ends = getActivity().findViewById(R.id.ends);
         timeEnds = getActivity().findViewById(R.id.endsText);
 
-        File file = getContext().getFileStreamPath("Countdown");
+        file = getContext().getFileStreamPath("Countdown");
 
         if (file.exists()) {
             if(!MainActivity.isTimerCreated) {
@@ -542,8 +555,9 @@ public class HomeFragment extends Fragment {
                     InputStreamReader inputStreamReader = new InputStreamReader(fileInputStream);
                     BufferedReader bufferedReader = new BufferedReader(inputStreamReader);
                     int parkingEndsMinutes = Integer.parseInt(bufferedReader.readLine());
+                    String parkingDate = bufferedReader.readLine();
 
-                    startParking(parkingEndsMinutes);
+                    startParking(parkingEndsMinutes, parkingDate);
 
                     if(MainActivity.isTimerCreated) {
 
@@ -591,8 +605,6 @@ public class HomeFragment extends Fragment {
                                     ends.setVisibility(View.VISIBLE);
                                     timeEnds.setVisibility(View.VISIBLE);
 
-                                    File file = getContext().getFileStreamPath("Countdown");
-
                                     if (file.exists()) {
                                         file.delete();
                                         MainActivity.countDownTimer.cancel();
@@ -602,13 +614,19 @@ public class HomeFragment extends Fragment {
 
                                     int parkingEndsMinutes = scan.nextInt() * 60 + scan.nextInt() % 60;
 
-                                    startParking(parkingEndsMinutes);
+                                    Date c = Calendar.getInstance().getTime();
+                                    SimpleDateFormat df = new SimpleDateFormat("yyyy-MM-dd");
+                                    String formattedDate = df.format(c);
+
+                                    startParking(parkingEndsMinutes, formattedDate);
 
                                     String fileName = "Countdown";
 
                                     try {
                                         FileOutputStream fileOutputStream = getActivity().openFileOutput(fileName, Context.MODE_APPEND);
                                         fileOutputStream.write(String.valueOf(parkingEndsMinutes).getBytes());
+                                        fileOutputStream.write("\n".getBytes());
+                                        fileOutputStream.write(formattedDate.getBytes());
                                         fileOutputStream.close();
                                     } catch (FileNotFoundException e) {
                                         e.printStackTrace();
@@ -629,17 +647,27 @@ public class HomeFragment extends Fragment {
         });
     }
 
-    public void startParking(int parkingEndsMinutes){
+    public long calculateTimeLeft(int parkingEndsMinutes, int hours, int minutes){
+
+        long timeLeftInMilliseconds = (parkingEndsMinutes - (hours * 60 + minutes)) * 60000;
+
+        return  timeLeftInMilliseconds;
+    }
+
+    public void startParking(int parkingEndsMinutes, String parkingDate){
 
         Calendar currentTime = GregorianCalendar.getInstance();
-
         currentTime.setTime(new Date());
+        timeLeftInMilliseconds = calculateTimeLeft(parkingEndsMinutes, currentTime.get(Calendar.HOUR_OF_DAY), currentTime.get(Calendar.MINUTE));
 
-        timeLeftInMilliseconds = (parkingEndsMinutes - (currentTime.get(Calendar.HOUR_OF_DAY) * 60 + currentTime.get(Calendar.MINUTE))) * 60000;
+        Date c = Calendar.getInstance().getTime();
+        SimpleDateFormat df = new SimpleDateFormat("yyyy-MM-dd");
+        String formattedDate = df.format(c);
 
-        if(timeLeftInMilliseconds <= 0){
-            File file = getContext().getFileStreamPath("Countdown");
-            file.delete();
+        if(timeLeftInMilliseconds <= 0 || parkingDate.compareTo(formattedDate) != 0){
+            if(file.exists()) {
+                file.delete();
+            }
             timeLeftInMilliseconds = -1;
             return;
         }
@@ -655,8 +683,9 @@ public class HomeFragment extends Fragment {
 
             @Override
             public void onFinish() {
-                File file = getContext().getFileStreamPath("Countdown");
-                file.delete();
+                if(file.exists()) {
+                    file.delete();
+                }
                 remaining.setVisibility(View.INVISIBLE);
                 timeLeft.setVisibility(View.INVISIBLE);
                 ends.setVisibility(View.INVISIBLE);
@@ -669,8 +698,8 @@ public class HomeFragment extends Fragment {
 
         int timeLeftInMinutes = (int) timeLeftInMilliseconds / 60000;
 
-        int hours = (int) timeLeftInMinutes / 60;
-        int minutes = (int) timeLeftInMinutes % 60;
+        int hours = timeLeftInMinutes / 60;
+        int minutes = timeLeftInMinutes % 60;
         int seconds = (int) timeLeftInMilliseconds % 60000 / 1000;
 
         String timeLeftText;
